Implementado AddressSchema con validación de datos

+++ backend/src/routes/addresses_route.py
@@ -1,22 +1,9 @@
 from flask import Blueprint, request, jsonify
 
-from ..utils.instantiations import ma, db
+from ..utils.instantiations import db
+from ..schemas.address_schema import AddressSchema
 from ..models.address_model import Address
 
-
-class AddressSchema(ma.Schema):
-    class Meta:
-        fields = (
-            "id",
-            "street",
-            "second_line_street",
-            "postal_code",
-            "city",
-            "address_user_id",
-        )
-
-
-address_schema = AddressSchema()
 addresses_schema = AddressSchema(many=True)
 
 addresses = Blueprint("addresses", __name__, url_prefix="/addresses")
@@ -25,8 +12,10 @@ addresses = Blueprint("addresses", __name__, url_prefix="/addresses")
 @addresses.route("/", methods=["POST"])
 def add_address():
     address_data = request.get_json()
+    context = {"mode": "create"}
+    address_schema = AddressSchema(load_instance=True, context=context)
 
-    new_address = Address(**address_data)
+    new_address = address_schema.load(address_data)
 
     db.session.add(new_address)
     db.session.commit()
@@ -42,9 +31,28 @@ def get_addresses():
     return addresses_schema.jsonify(all_addresses), 200
 
 
-@addresses.route("/<address_id>", methods=["GET", "DELETE", "PUT"])
+@addresses.route("/<address_id>", methods=["GET", "PUT", "DELETE"])
 def handle_address(address_id):
     address = Address.query.get(address_id)
+    address_schema = AddressSchema()
+
+    if request.method == "PUT":
+        address_data = request.get_json()
+
+        context = {
+            "mode": "update",
+            "expected_id": address_id,
+            "expected_address_user_id": address.address_user_id,
+        }
+        address_schema.context = context
+
+        address_schema.load(address_data)
+
+        for key, value in address_data.items():
+            setattr(address, key, value)
+
+        db.session.commit()
+        return address_schema.jsonify(address), 200
 
     if request.method == "DELETE":
         db.session.delete(address)
@@ -52,22 +60,10 @@ def handle_address(address_id):
 
         return jsonify(msg="La dirección ha sido eliminada correctamente"), 200
 
-    if request.method == "PUT":
-        address_data = request.get_json()
-
-        # TODO: Falta filtrar resultados
-        address.street = address_data["street"]
-        address.second_line_street = address_data["second_line_street"]
-        address.postal_code = address_data["postal_code"]
-        address.city = address_data["city"]
-
-        db.session.commit()
-        return address_schema.jsonify(address), 200
-
     return address_schema.jsonify(address), 200
 
 
 @addresses.route("/users/<user_id>", methods=["GET"])
 def get_addresses_user(user_id):
-    address = Address.query.filter_by(address_user_id=user_id).all()
-    return address_schema.jsonify(address), 200
+    user_addresses = Address.query.filter_by(address_user_id=user_id).all()
+    return addresses_schema.jsonify(user_addresses), 200
